Fix security issues (Refs #2225)

+++ applications/routing/yours/branches/version-1.0-via/www/killall.php
@@ -1,40 +1,50 @@
 <?php
+$allowed_hosts = array('213.10.112.251');
+
 $ps = array();
 $output = array();
+$host = $_SERVER["REMOTE_ADDR"];
 
 echo '<html xmlns="http://www.w3.org/1999/xhtml" xml:lang="en">';
 echo '<head></head><body>';
 
-echo "Start killing Gosmore processes:<br><br>";
-
-exec("ps ax | grep gosmore", $ps, $return_var);
-
-echo "Found ".Count($ps)." processes containing 'Gosmore'<br>";
-//print_r($ps);
-foreach ($ps as $row => $process)
+if (in_array($host, $allowed_hosts) == false )
 {
-	echo "$process<br>";
+	echo "You are not allowed to use this function ($host)";
 }
-echo "<br>Actions:<br>";
-foreach ($ps as $row => $process)
+else
 {
-	$bKill = true;
-	$properties = array();
-	$properties = split(" ", $process);
-	
-	foreach ($properties as $item => $property)
+	echo "Start killing Gosmore processes:<br><br>";
+
+	exec("ps ax | grep gosmore", $ps, $return_var);
+
+	echo "Found ".Count($ps)." processes containing 'gosmore'<br>";
+	//print_r($ps);
+	foreach ($ps as $row => $process)
 	{
-		//echo "property ".$item." = ".$property."\n";
-		if (strstr($property, "grep"))
-		{
-			echo "Not killing ".$properties[0]."<br>";
-			$bKill = false;
-		}
+		echo "$process<br>";
 	}
-	if ($bKill == true)
+	echo "<br>Actions:<br>";
+	foreach ($ps as $row => $process)
 	{
-		echo "Killing ".$properties[0]."<br>";
-		exec("kill -9 ".$properties[1], $output);
+		$bKill = true;
+		$properties = array();
+		$properties = split(" ", $process);
+	
+		foreach ($properties as $item => $property)
+		{
+			//echo "property ".$item." = ".$property."\n";
+			if (strstr($property, "grep"))
+			{
+				echo "Not killing ".$properties[0]."<br>";
+				$bKill = false;
+			}
+		}
+		if ($bKill == true)
+		{
+			echo "Killing ".$properties[0]."<br>";
+			exec("kill -9 ".$properties[1], $output);
+		}
 	}
 }
 echo '</body>';
